% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-ltmv.R
\name{finn_storrelse_diag}
\alias{finn_storrelse_diag}
\title{Hent ut antall, navn, eller prosent for diagnosegrupper av ulik størrelse}
\usage{
finn_storrelse_diag(d_n_diaggruppe_akt, alderkat, type, str_orden)
}
\arguments{
\item{d_n_diaggruppe_akt}{Tibble/dataramme med diagnosegrupper, alderskategori, antall og andel.}

\item{alderkat}{Tekstvektor med alderskategorien skal det sjekkes for, "Barn" eler "Voksen".}

\item{type}{Tekstvektor med navnet på kolonnen som utverdien skal hentes fra.}

\item{str_orden}{Heltall som sier hvilken diagnosegruppe det skal hentes tall for.
Henter for den n-te største diagnosegruppen etter antall.}
}
\value{
Verdien i kolonnen \code{type} for den \code{str_orden}-te største diagnosegruppen
etter antall i \code{d_n_diaggruppe_akt},
filtrert på alderskategori \code{alderkat}.
}
\description{
Hent ut antall, navn, eller prosent for diagnosegrupper av ulik størrelse
}
