% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superbreitt_format.R
\name{fjern_duplikate_skjema}
\alias{fjern_duplikate_skjema}
\title{Sjekk for duplikat oppfølingsskjema}
\usage{
fjern_duplikate_skjema(d, aar)
}
\arguments{
\item{d}{Skjemaet ventfol.}

\item{aar}{Tekstvariabel. Året for oppfølgning (f.eks. "1")}
}
\value{
Datasettet d, men der eventuelle duplikate oppfølgingsskjema er fjerne.
Om det ikke finnes duplikate oppfølgingsskjema blir datasettet
retunert uendret.
}
\description{
Funksjonen tar inn datasettet ventfol og aar for oppfølging,
og beholder bare det øverste skjema dersom en pasient ligger
inne med flere skjema for oppfølging 1 år eller 3 år.
}
\details{
Funksjonen gir varsel dersom det finnes duplikate skjema.
}
\examples{
\dontrun{
les_data_ltmv()

d_mce = select(d_full_mce, mceid, patient_id, parent_mce)

d_full_ventfol_ekstra = left_join(
  x = d_full_ventfol,
  y = d_mce,
  by = join_by(mceid == mceid)
)

fjern_duplikate_skjema(d_full_ventfol_ekstra, "3")
}
}
