% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_antall_skjema_tabell.R
\name{formater_antall_skjema_tabell}
\alias{formater_antall_skjema_tabell}
\title{Formater antall skjema-tabell}
\usage{
formater_antall_skjema_tabell(d_antall_skjema)
}
\arguments{
\item{d_antall_skjema}{Dataramme frå \code{\link[=aggreger_antall_skjema_tabell]{aggreger_antall_skjema_tabell()}}.}
}
\value{
\code{d_antall_skjema} formatert som HTML-tabell,
med samlande overskrifter, hover-effekt og så vidare.
}
\description{
Tek inn ei dataramme med tal på ferdige og uferdige skjema,
frå \code{\link[=aggreger_antall_skjema_tabell]{aggreger_antall_skjema_tabell()}},
og gjev den ut som HTML-tabell klar til å visast i Rapporteket.
}
\examples{
d_antall_skjema = ltmv:::aggreger_antall_skjema_tabell(
  Sys.Date() - 365, Sys.Date()
)
ltmv:::formater_antall_skjema_tabell(d_antall_skjema)
}
\keyword{internal}
