% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer-ltmv.R
\name{ki_blodgass}
\alias{ki_blodgass}
\title{Kvalitetsindikator for blodgass}
\usage{
ki_blodgass(d_ventreg)
}
\arguments{
\item{d_ventreg}{Basisskjema.}
}
\value{
KI-datasett egnet for bruk med \code{\link[rapwhale:aggreger_ki_prop]{rapwhale::aggreger_ki_prop()}}.
}
\description{
Lag KI-datasett for andel forløp som fikk utført blodgassmåling ved start.
}
