% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_elektiv_behandling.R
\name{ki_elektiv}
\alias{ki_elektiv}
\title{Kvalitetsindikator for elektiv behandling}
\usage{
ki_elektiv(d_ventreg)
}
\arguments{
\item{d_ventreg}{Basisskjema.}
}
\value{
KI-datasett egnet for bruk med \code{\link[rapwhale:aggreger_ki_prop]{rapwhale::aggreger_ki_prop()}}.
}
\description{
Lag KI-datasett for andel forløp med elektiv (planlagt) behandlingsstart.
}
