% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer-ltmv.R
\name{ki_prom}
\alias{ki_prom}
\title{Kvalitetsindikator for PROM}
\usage{
ki_prom(d_prom)
}
\arguments{
\item{d_prom}{PROM-datasett på langt format. Alle PROM-spørsmålene
må være i én kolonne med tilhørende verdier i en annen.
Datasettet må også inneholde en kolonne med navn "mangler_ikke"
som har verdien 0 for spørsmål som ikke er besvart og verdien
1 for spørsmål som er besvart.}
}
\value{
KI-datasett egnet for bruk med \code{\link[rapwhale:aggreger_ki_prop]{rapwhale::aggreger_ki_prop()}}.
}
\description{
Lag KI-datasett for andel pasienter som har besvart PROM ved behandlingsstart.
}
