% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_aktivitetsoversikt.R
\name{lag_aktivitetsoversikt}
\alias{lag_aktivitetsoversikt}
\title{Lag HTML-tabell med aktivitetsoversikt}
\usage{
lag_aktivitetsoversikt(d)
}
\value{
HTML-tabell med aktivitetsoversikt per dags dato i registeret.
}
\description{
Lag HTML-tabell med aktivitetsoversikt
}
