% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_antall_skjema_tabell.R
\name{lag_antall_skjema_tabell}
\alias{lag_antall_skjema_tabell}
\title{Lag HTML-tabell med oversikt over ferdige og uferdige skjema på sjukehusnivå}
\usage{
lag_antall_skjema_tabell(fra, til)
}
\arguments{
\item{fra}{Dato, eller eventuelt tekst på formatet "yyyy-mm-dd".
Skjema oppretta frå og med denne datoen vert inkludert.}

\item{til}{Dato, eller eventuelt tekst på formatet "yyyy-mm-dd".
Skjema oppretta til og med denne datoen vert inkludert.}
}
\value{
HTML-tabell med talet på ferdige og uferdige skjema på sjukehusnivå
for registrerings-, avslutnings- og ulike oppfylgjingsskjema.
}
\description{
Funksjonen tek inn ein dato \code{fra} og ein dato \code{til},
og gjev ut ein HTML-tabell med talet på ferdige og uferdige skjema på
sjukehusnivå for registrerings-, avslutnings- og ulike oppfylgjingsskjema.
}
\examples{
lag_antall_skjema_tabell(Sys.Date() - 365, Sys.Date())
}
