% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_datasett_dashboard.R
\name{lag_datasett_dashboard}
\alias{lag_datasett_dashboard}
\title{Lag dashboard-datasett}
\usage{
lag_datasett_dashboard(
  fra,
  til,
  alderkategori,
  alderkategori_naa,
  kjonn,
  inkluder_missing,
  resh_id,
  user_role,
  enhetstype,
  per_rhf,
  per_hf,
  per_sykehus
)
}
\arguments{
\item{fra}{Dato. Skjema frå og med denne datoen skal inkluderast.}

\item{til}{Dato. Skjema til og med denne datoen skal inkluderast.}

\item{alderkategori}{Tekstvektor for alderkategori ved start. Éin eller fleire av verdiane "barn", "voksen" og NA_character.}

\item{alderkategori_naa}{Tekstvektor for alderkategori nå. Éin eller fleire av verdiane "barn", "voksen" og NA_character.}

\item{kjonn}{Numerisk vektor. Éin eller fleire av verdiane 1, 2, NA_real_.
1 er mann, 2 er kvinne.}

\item{inkluder_missing}{Logisk variabel. \code{TRUE} eller \code{FALSE} alt etter om skjema med manglande
startdato skal inkluderast.}

\item{resh_id}{Tekstvektor med RESH-ID datasettet eventuelt skal filtrerast på.}

\item{user_role}{Tekstvektor med brukarrolle som avgjer om datasettet skal filtrerast eller
ikkje.}

\item{enhetstype}{Tekstvektor med enhetstypene ("RHF", "HF", eller "Sykehus") data kan filtreres
ut mot. Hentes fra \code{shiny::radioButtons()} med \code{id = enhet_type} i app_server.R.}

\item{per_rhf}{Teksvektor med de ulike RHF-ene data kan filtreres ut mot.  Hentes fra
\code{shiny::checkboxGroupInput()} med \code{id = rhf_utvalg_dashboard} i app_server.R.}

\item{per_hf}{Teksvektor med de ulike HF-ene data kan filtreres ut mot.  Hentes fra
\code{shiny::selectInput()} med \code{id = hf_utvalg_dashboard} i app_server.R.}

\item{per_sykehus}{Teksvektor med de ulike sykehusene data kan filtreres ut mot.  Hentes fra
\code{shiny::selectInput()} med \code{id = sykehus_utvalg_dashboard} i app_server.R.}
}
\value{
Skjemaet ventreg med ekstra informasjon per rad,
til bruk i dashboardvisinga.
}
\description{
Funksjonen les inn skjemaet ventreg, legg til diverse infomasjon
til kvar rad, og gjer datasettet klart til bruk i dashboardet.
}
\examples{
\dontrun{
lag_datasett_dashboard(
  fra = "2020-01-01",
  til = "2023-01-01",
  alderkategori = "barn",
  alderkategori_naa = "voksen",
  kjonn = c(1, 2, NA_real_),
  inkluder_missing = TRUE,
  resh_id = 99999,
  user_role = "SC"
)
}
}
