% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_fig_aldersfordeling.R
\name{lag_fig_aldersfordeling}
\alias{lag_fig_aldersfordeling}
\title{Lag figur med aldersfordeling}
\usage{
lag_fig_aldersfordeling(d)
}
\arguments{
\item{d}{Dataramme som må innehalda variabelen \code{alder}.}
}
\value{
Histogram med aldersfordeling.
}
\description{
Tek inn ei dataramme som minst inneheld variabelen \code{alder},
og lagar ein figur med aldersfordeling.
}
\examples{
\dontrun{
hent_skjema("ventreg") |>
  legg_til_pasientid(mceid) |>
  legg_til_pasientinfo(patient_id) |>
  legg_til_alder_og_kategori(birth_date, start_date) |>
  lag_fig_aldersfordeling()
}
}
