% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_fig_diagnosefordeling.R
\name{lag_fig_diagnosefordeling}
\alias{lag_fig_diagnosefordeling}
\title{Lag figur med diagnosefordeling}
\usage{
lag_fig_diagnosefordeling(d)
}
\description{
Lag figur med diagnosefordeling
}
