% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legg-til-info.R
\name{legg_til_aktiv_behandling}
\alias{legg_til_aktiv_behandling}
\title{Legg til info aktiv behandling}
\usage{
legg_til_aktiv_behandling(d)
}
\arguments{
\item{d}{Dataramme.
Må inneholde kolonnene \code{deceased}, \code{stop_date} og \code{conclude_status}.}
}
\value{
Datarammen \code{d}, med en ekstra kolonne \code{aktiv_behandling} som er
\code{TRUE} eller \code{FALSE} avhengig av om forløpet er pågående eller avsluttet.
}
\description{
Tar inn en dataramme,
og legger til en ekstra kolonne \code{aktiv_behandling} som er
\code{TRUE} eller \code{FALSE} avhengig av om forløpet er pågående eller avsluttet.
}
\details{
Et forløp regnes som pågående/aktivt hvis det verken finnes en stoppdato,
registrerte opplysninger om at pasienten er død,
eller et ferdigstilt tilhørende avslutningsskjema.
}
\examples{
\dontrun{
hent_skjema("ventreg") |>
  legg_til_pasientid(mceid) |>
  legg_til_pasientinfo(patient_id) |>
  legg_til_stoppinfo(mceid) |>
  legg_til_aktiv_behandling()
}
}
