% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legg-til-info.R
\name{legg_til_alder_og_kategori}
\alias{legg_til_alder_og_kategori}
\title{Legg til alder og alderskategori}
\usage{
legg_til_alder_og_kategori(d, fodselsdato_varnavn, hendelsesdato_varnavn)
}
\arguments{
\item{d}{Dataramme.
Må inneholde kolonner med navnene gitt i
\code{fodselsdato_varnavn} og \code{hendelsesdato_varnavn}.}

\item{fodselsdato_varnavn}{Navn på kolonnen i \code{d} som inneholder fødselsdato.}

\item{hendelsesdato_varnavn}{Navn på kolonnen i \code{d} som inneholder dato
som skal brukes til å beregne alder.}
}
\value{
Datarammen \code{d}, med ekstra kolonner med alder ved datoen i kolonnen
\code{hendelsesdato_varnavn}, og tilhørende alderskategori barn/voksen
(over/under 18 år).
}
\description{
Tar inn en dataramme,
og legger til ekstra kolonner med alder og alderskategori barn/voksen
(over/under 18 år).
}
\examples{
\dontrun{
hent_skjema("ventreg") \%>\%
  legg_til_pasientid(mceid) \%>\%
  legg_til_pasientinfo(patient_id) \%>\%
  legg_til_alder_og_kategori(birth_date, start_date)
}
}
