% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-ltmv.R
\name{legg_til_fylke_rhf_og_resh_info}
\alias{legg_til_fylke_rhf_og_resh_info}
\title{Legg til oppdaterte fylkesnavn og rekkefølge på helseregioner}
\usage{
legg_til_fylke_rhf_og_resh_info(d)
}
\arguments{
\item{d}{Tibble/dataramme med data fra et av skjemaene ventreg,
ventfol eller conclude,
f.eks. hentet med \code{\link[=les_data_ltmv]{les_data_ltmv()}}.}
}
\value{
Inndatarammen \code{d} med ekstra kolonner med info om fylke og RHF lagt til.
}
\description{
Legg til oppdaterte fylkesnavn og rekkefølge på helseregioner
}
