% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-ltmv.R
\name{legg_til_hf_rhf_navn}
\alias{legg_til_hf_rhf_navn}
\title{Legg til HF-navn og RHF-navn}
\usage{
legg_til_hf_rhf_navn(d)
}
\arguments{
\item{d}{Tibble/dataramme med data fra et av skjemaene ventreg,
ventfol eller conclude,
f.eks. hentet med \code{\link[=les_data_ltmv]{les_data_ltmv()}}.}
}
\value{
Inndatarammen \code{d} med ekstra kolonner med info om HF lagt til.
}
\description{
Legg til HF-navn og RHF-navn
}
