% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-ltmv.R
\name{legg_til_oppdaterte_fylker_og_rekkefolge_helseregion}
\alias{legg_til_oppdaterte_fylker_og_rekkefolge_helseregion}
\title{Legg til oppdaterte fylkesnavn og rekkefølge på helseregioner}
\usage{
legg_til_oppdaterte_fylker_og_rekkefolge_helseregion(d)
}
\arguments{
\item{d}{Tibble/dataramme med data fra et av skjemaene ventreg,
ventfol eller conclude,
f.eks. hentet med \code{\link[=les_data_ltmv]{les_data_ltmv()}}.}
}
\value{
Inndatarammen \code{d} med ekstra kolonner med info om fylke og RHF lagt til.
}
\description{
Legg til oppdaterte fylkesnavn og rekkefølge på helseregioner
}
