% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-ltmv.R
\name{legg_til_overordnet_diag}
\alias{legg_til_overordnet_diag}
\title{Legg til overordnede diagnosegrupper}
\usage{
legg_til_overordnet_diag(d_ventreg, grunnmappe)
}
\arguments{
\item{d_ventreg}{Tibble/dataramme med data fra skjemaet ventreg,
f.eks. hentet med \code{\link[=les_data_ltmv]{les_data_ltmv()}}.}

\item{grunnmappe}{Tekststreng med adresse til ltmv sin kvalitetsserver.}
}
\value{
Inndatarammen \code{d_ventreg} med ekstra kolonner med info om overordnede
diagnosegrupper lagt til.
}
\description{
Legg til overordnede diagnosegrupper
}
