% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legg-til-info.R
\name{legg_til_pasientid}
\alias{legg_til_pasientid}
\title{Legg til pasientid}
\usage{
legg_til_pasientid(d, skjemaid_varnavn)
}
\arguments{
\item{d}{Dataramme. Må inneholde en kolonne med navnet gitt i \code{skjemaid_varnavn}.}

\item{skjemaid_varnavn}{Navn på kolonnen i \code{d} som inneholder skjema-ID.}
}
\value{
Datarammen \code{d}, med en ekstra kolonne med med pasient-ID.
}
\description{
Tar inn en dataramme, og legger til en ekstra kolonne med pasient-ID.
}
\examples{
\dontrun{
hent_skjema("ventreg") \%>\%
  legg_til_pasientid(mceid)
}
}
