% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legg-til-info.R
\name{legg_til_pasientinfo}
\alias{legg_til_pasientinfo}
\title{Legg til pasientinfo}
\usage{
legg_til_pasientinfo(d, pasientid_varnavn)
}
\arguments{
\item{d}{Dataramme. Må inneholde en kolonne med navnet gitt i \code{pasientid_varnavn}.}

\item{pasientid_varnavn}{Navn på kolonnen i \code{d} som inneholder pasient-ID.}
}
\value{
Datarammen \code{d}, med ekstra kolonner med info om pasienten.
}
\description{
Tar inn en dataramme, og legger til ekstra kolonner med info om pasienten.
}
\details{
Funksjonen knytter datarammen \code{d} til pasientskjema basert på pasient-IDer
i kolonnen \code{pasientid_varnavn}.
Fra pasientskjema hentes dato for registrering, fødselsdato, kjønn,
info om pasienten er død eller levende, og eventuell dødsdato.
Denne infoen legges så til som nye kolonner i den originale datarammen.
}
\examples{
\dontrun{
hent_skjema("ventreg") |>
  legg_til_pasientid(mceid) |>
  legg_til_pasientinfo(patient_id)
}
}
