% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legg-til-info.R
\name{legg_til_stoppinfo}
\alias{legg_til_stoppinfo}
\title{Legg til info om behandlingsstopp}
\usage{
legg_til_stoppinfo(d, skjemaid_varnavn)
}
\arguments{
\item{d}{Dataramme. Må inneholde en kolonne med navnet gitt i \code{skjemaid_varnavn}.}

\item{skjemaid_varnavn}{Navn på kolonnen i \code{d} som inneholder skjema-ID.}
}
\value{
Datarammen \code{d}, med ekstra kolonner med info om behandlingsstopp.
}
\description{
Tar inn en dataramme,
og legger til ekstra kolonner med info om behandlingsstopp.
}
\examples{
\dontrun{
hent_skjema("ventreg") \%>\%
  legg_til_stoppinfo(mceid)
}
}
