% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_data_ltmv.R
\name{les_data_ltmv}
\alias{les_data_ltmv}
\title{Les inn LTMV-data}
\usage{
les_data_ltmv(
  mappe_dd,
  dato = NULL,
  maksdato = NULL,
  status = 1,
  valider = TRUE,
  omgjevnad = .GlobalEnv
)
}
\arguments{
\item{mappe_dd}{Grunnmappe som datadumpar skal hentast frå.
Må ha undermapper med datonamn (format: \code{ÅÅÅÅ-MM-DD}).}

\item{dato}{Datoen for den aktuelle datadumpen (anten tekst på formatet \code{ÅÅÅÅ-MM-DD}
eller eit \link[base:Date]{base::Date}-objekt). Vert brukt til å velja rett undermappe
til \code{mappe_dd}. Viss \code{NULL} (standard), vert nyaste dato brukt.}

\item{maksdato}{Dato for siste startdato,
oppfølgingsdato og avslutningsdato som skal vera med i datauttrekket
(same datoformat som for \code{dato}).
Oppføringar som tilhøyrer seinare datoar,
vert filtrerte ut frå skjemaa.
Viss \code{NULL} (standard), vert det ikkje noko filtrering på datoane.}

\item{status}{Type oppføringar som skal lesast.
Sjå \code{\link[rapwhale:les_dd_oqr]{rapwhale::les_dd_oqr()}} for informasjon om moglege verdiar.
Som standard vert berre ferdigstilte oppføringar lesne.}

\item{valider}{Skal dataa validerast,
dvs. sjekkast for feilverdiar før innlesing? (Standard: ja.)
Omfattar testar ved bruk av kodeboka.}

\item{omgjevnad}{Omgjevnad som dataobjekta skal lagrast til.
Standard er den globale omgjevnaden.}
}
\description{
Les inn LTMV-datadumpar og lagra dei som objekt,
eventuelt med automatisk validering.
}
\details{
Alle standard datafiler til registeret vert lesne inn, eventuelt validerte,
filtrerte (sjå nedanfor) og så lagra som R-objekt.
Dei får namna \code{d_full_skjemanamn},
der \code{skjemanamn} er namnet på det aktuelle skjemaet
(for eksempel \code{d_full_ventfol} for oppfølgingskjemaet).
Som standard vert dei lagra i den globale omgjevnaden,
og dei vil overskriva eventuelle eksisterande objekt med same namn.

I tillegg til alle standardskjemaa som er nemnde i kodeboka,
vert datasett for \code{mce}- og \code{patientlist}-filene òg lagra.

Kodeboka for registeret vert òg lagra, med namnet \code{kb}.
}
\examples{
\dontrun{
# Les inn ferdigstilte skjema (nyaste datadumpar)
les_data_ltmv()
}
}
