% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_data}
\alias{query_data}
\alias{query_all_hospitals}
\title{Get data from LTMV}
\usage{
query_all_hospitals(registry_name, resh_id, ...)
}
\arguments{
\item{registry_name}{Character string with name of the registry from which
data are to be queried}

\item{resh_id}{Integer providing the organization id. Useful for data
filtering}

\item{...}{Optional arguments to pass to the function. If \code{session} is
provided this will be assumed a valid shiny session object and hence logging
may be performed}
}
\value{
Data frame of registry data
}
\description{
Functions to query data from a database holding LTMV data at Rapporteket.
Providing these function with the shiny session object logging of queries
may also be performed
}
