% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn_ut.R
\name{regn_andel_missing}
\alias{regn_andel_missing}
\title{Finn andel som har en m책ling for sentrale variabler brukt i 책rsrapporten}
\usage{
regn_andel_missing(d, var)
}
\arguments{
\item{d}{Tibble/dataramme.}

\item{var}{Kolonne det skal regnes antall og andel missing for.}
}
\value{
Tibble/dataramme med 3 kolonner og 1 rad.
Kolonnene er \code{teller} (antall rader i \code{d} som har en verdi i kolonnen \code{var}),
\code{nevner} (antall rader i \code{d}) og \code{prop}
(andel rader i \code{d} som har en verdi i kolonnen \code{var}(\code{teller/nevner})).
}
\description{
Finn andel som har en m책ling for sentrale variabler brukt i 책rsrapporten
}
