% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn_ut.R
\name{regn_antall}
\alias{regn_antall}
\title{Regn ut antall og prosent for diagnosegrupper}
\usage{
regn_antall(d, alderkat = TRUE)
}
\arguments{
\item{d}{Tibble/dataramme med data fra skjemae ventreg,
f.eks. hentet med \code{\link[=les_data_ltmv]{les_data_ltmv()}},
med info om diagnosegruppe lagt til med \code{\link[=legg_til_overordnet_diag]{legg_til_overordnet_diag()}},
og en kolonne \code{alderkat} som sier om pasienten er barn (under 18 år)
eller voksen (18 år eller eldre).}

\item{alderkat}{Boolsk variabel som sier om det skal gupperes på kolonnen \code{alderkat} i \code{d}.
Standardverdi er \code{TRUE}.}
}
\value{
Tibble med oppsumert antall og andel pasienter de ulike diagnosegruppene,
eventuelt gruppert på alderskategori (barn/voksen) også.
}
\description{
Regn ut antall og prosent for diagnosegrupper
}
