% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn_ut.R
\name{regn_ut_gj_trakestomi}
\alias{regn_ut_gj_trakestomi}
\title{Regn ut prosent for andel trakeostomi eller maske (for bruk i teksten)}
\usage{
regn_ut_gj_trakestomi(d, trakeostomi_type)
}
\arguments{
\item{d}{Tibble/dataramme med data fra et av skjemaene ventreg eller ventfol,
f.eks. hentet med \code{\link[=les_data_ltmv]{les_data_ltmv()}}.}

\item{trakeostomi_type}{Hvilken type ventilering skal det regnes ut andel for?
Verdien 1 for trakeostomi, verdien 2 for non-invasiv.}
}
\value{
Andel forløp i \code{d} med ventilering av typen \code{trakeostomi_type}.
(I prosent, med én desimal og uten prosenttegn, altså andel * 100)
}
\description{
Regn ut prosent for andel trakeostomi eller maske (for bruk i teksten)
}
