% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superbreitt_format.R
\name{superbreitt_format}
\alias{superbreitt_format}
\title{Lag datasett på superbredt format}
\usage{
superbreitt_format(
  d_full_patientlist,
  d_full_mce,
  d_full_ventreg,
  d_full_ventfol,
  d_full_conclude
)
}
\arguments{
\item{d_full_patientlist}{Skjemaet patientlist.}

\item{d_full_mce}{Skjemaet mce.}

\item{d_full_ventreg}{Skjemaet ventreg.}

\item{d_full_ventfol}{Skjemaet ventfol.}

\item{d_full_conclude}{Skjemaet conclude.}
}
\value{
Datasettet d_superbrei som inneholder alle skjemaene på
superbredt format.
}
\description{
Funksjonen tar inn skjemaene patientlist, mce, ventreg, ventfol,
og conclude, og legger på en bokstav i starten av hver variabel
før skjemaene blir slått sammen til datasettet d_superbrei.
}
\details{
Funksjonen tar i bruk fjern_duplikate_skjema() for ventfol, som
beholder bare det øverste skjema dersom en pasient ligger inne
med flere skjema for oppfølging 1 år, eller 3 år.

"p_" blir lagt på starten av variablene i patientlist.
"r_" blir lagt på starten av variablene i ventreg.
"f1_", "f3", "fah_" og "lf_" blir lagt på starten av variablene
i ventfol avhengig om variablene er tilknyttet oppfølgning 1 år,
3 år, ad hoc, eller siste oppfølging.
}
\examples{
\dontrun{
les_data_ltmv()

superbreitt_format(
  d_full_patientlist,
  d_full_mce,
  d_full_ventreg,
  d_full_ventfol,
  d_full_conclude
)
}
}
