% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_antall_skjema_tabell.R
\name{aggreger_antall_skjema_tabell}
\alias{aggreger_antall_skjema_tabell}
\title{Lag dataramme med oversikt over ferdige og uferdige skjema på sjukehusnivå}
\usage{
aggreger_antall_skjema_tabell(d_skjemaoversikt)
}
\arguments{
\item{d_skjemaoversikt}{Skjemaet «SkjemaOversikt» frå LTMV-databasen, som ei dataramme.
Eventuelt filtrert på t.d. sjukehus.}
}
\value{
Dataramme med talet på ferdige og uferdige skjema på sjukehusnivå
for registrerings-, avslutnings- og ulike oppfylgjingsskjema.
}
\description{
Funksjonen tek inn ei dataramme \code{d_skjemaoversikt},
og gjev ut ei dataramme med talet på ferdige og uferdige skjema på
sjukehusnivå for registrerings-, avslutnings- og ulike oppfylgjingsskjema.
}
\examples{
d_skjemaoversikt = hent_skjema("SkjemaOversikt")
ltmv:::aggreger_antall_skjema_tabell(d_skjemaoversikt, Sys.Date() - 365, Sys.Date())
}
\keyword{internal}
