% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent_skjema.R
\name{hent_skjema}
\alias{hent_skjema}
\title{Hent skjema}
\usage{
hent_skjema(skjemanamn, registernamn = "ltmv")
}
\arguments{
\item{skjemanamn}{Tekststreng med namn på skjemaet som skal hentast.}

\item{registernamn}{Tekststreng med namn på register. Standardverdi er "ltmv".
Argumentet er med for å kunna køyra testar på testdatabase.}
}
\value{
Tibble med all data frå skjemaet med namn \code{skjemanamn}.
}
\description{
Hentar all data frå eitt skjema.
}
\details{
Fuksjonen brukar \code{\link[rapbase:loadRegData]{rapbase::loadRegData()}}
til å henta all data frå skjemaet med namn \code{skjemanamn}:
\verb{SELECT * FROM skjemanamn}

I databasen er det brukt store bokstavar i variabelnamna.
I utdata frå denne funksjonen er dei gjort om til små bokstavar med
\code{\link[stringr:case]{stringr::str_to_lower()}}.
}
\examples{
\dontrun{
hent_skjema("ventreg")
}
}
