% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_blodgass_forste_aar.R
\name{ki_blodgass_forste_aar}
\alias{ki_blodgass_forste_aar}
\title{Kvalitetsindikator for oppfølging av blodgass første år.}
\usage{
ki_blodgass_forste_aar(d_full_reg_forste_aar_ahoc, rapporteringsdato)
}
\arguments{
\item{d_full_reg_forste_aar_ahoc}{Tar inn et datasett på
superbredt"-format som inneholder ventreg, ventfol
(filtrert for year == 1 og year == -1 (første adhoc))
og conclude (one-to-one).
Variabelnavnene i ventfol knyttet til year == 1 må i
tillegg starte med "f1_" (f.eks. "f1_qlwakeup") og
variabelnavnene i ventfol knyttet til year == -1 må
starte med "fah_" (f.eks. "fah_po2_air").}

\item{rapporteringsdato}{Tar inn en dato (År-Måned-Dag)
som er den siste datoen som blir tatt med i beregningen
(innen to år). F.eks. dersom datoen er "2023-12-31",
blir siste år med komplett data 2021.}
}
\value{
KI-datasett egnet for bruk med \code{\link[rapwhale:aggreger_ki_prop]{rapwhale::aggreger_ki_prop()}}.
}
\description{
Lag KI-datasett for andel forløp som fikk utført
blodgassmåling ved oppfølging - innen 2 år.
Dersom pasienten har avsluttet eller dødd innen
2 år uten å ha besvart PROM, blir pasienten ikke
tatt med i beregningen av nevner eller teller.

Tar inn et datasett på "superbredt"-format som
inneholder ventreg, ventfol (filtrert på year == 1)
og conclude.
Variabelnavnene i ventfol må i tillegg starte med
"f1_" (f.eks. "f1_qlwakeup").
}
