% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_antall_skjema_tabell.R
\name{lag_antall_skjema_tabell}
\alias{lag_antall_skjema_tabell}
\title{Lag HTML-tabell med oversikt over ferdige og uferdige skjema på sjukehusnivå}
\usage{
lag_antall_skjema_tabell(
  fra,
  til,
  alderkategori,
  aktiv_behandling,
  resh_id,
  user_role,
  vis_hf,
  vis_rhf,
  rhf_utvalg
)
}
\arguments{
\item{fra}{Dato, eller eventuelt tekst på formatet "yyyy-mm-dd".
Skjema oppretta frå og med denne datoen vert inkludert.}

\item{til}{Dato, eller eventuelt tekst på formatet "yyyy-mm-dd".
Skjema oppretta til og med denne datoen vert inkludert.}

\item{alderkategori}{Tekstvektor med éin eller fleire av verdiane "barn", "voksen" og "".}

\item{aktiv_behandling}{Logisk vektor med éin eller begge verdiane \code{TRUE} og \code{FALSE}.}

\item{resh_id}{Tekststreng med RESH-ID til eininga til innlogga brukar.}

\item{user_role}{Tekststreng med rolla til innlogga brukar.
Til dømes "SC" (system coordinator) eller "LU" (local user).}

\item{vis_hf}{Viser til \code{checkboxInput()} med etiketten "kun_hf" i ui som skal vise sykehusene gruppert per HF}

\item{vis_rhf}{Viser til \code{checkboxInput()} med etiketten "kun_rhf" i ui som skal vise data kun for RHF-ene}

\item{rhf_utvalg}{Viser til \code{checkboxGroupInput()} med etiketten "rhf_utvalgt" i ui som skal si
hvilke RHF-er det vises data for.}
}
\value{
HTML-tabell med talet på ferdige og uferdige skjema på sjukehusnivå
for registrerings-, avslutnings- og ulike oppfylgjingsskjema.
}
\description{
Funksjonen tek inn datoar \code{fra} og \code{til},
tekstvektorar \code{alderkategori}, \code{resh_id} og \code{user_role},
og ein logisk vektor \code{aktiv_behandling},
og gjev ut ein HTML-tabell med talet på ferdige og uferdige skjema på
sjukehusnivå for registrerings-, avslutnings- og ulike oppfylgjingsskjema.
}
\details{
Funksjonen tek inn datoar \code{fra} og \code{til},
tekstvektorar \code{alderkategori}, \code{resh_id} og \code{user_role},
og ein logisk vektor \code{aktiv_behandling},
og gjev ut ein HTML-tabell med talet på ferdige og uferdige skjema på
sjukehusnivå for registrerings-, avslutnings- og ulike oppfylgjingsskjema.

Viss \code{user_role} ikkje er lik "SC" (system coordinator),
vert berre data frå eininga med RESH-ID \code{resh_id} vist.
}
\examples{
\dontrun{
lag_antall_skjema_tabell(
  fra = Sys.Date() - 365,
  til = Sys.Date(),
  alderkategori = "voksen",
  aktiv_behandling = TRUE,
  resh_id = 99999,
  user_role = "SC",
  vis_hf = FALSE,
  vis_rhf = FALSE,
  rhf_utvalg = c(
    "HELSE SØR-ØST RHF", "HELSE VEST RHF", "HELSE MIDT-NORGE RHF",
    "HELSE NORD RHF", "PRIVAT/IDEELL ORGANISASJON"
  )
)
}
}
