% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_spcfigur_ki_blodgass.R
\name{lag_spcfigur_ki_blodgass}
\alias{lag_spcfigur_ki_blodgass}
\title{Lag SPC-figur for blodgassindikator}
\usage{
lag_spcfigur_ki_blodgass(d_ventreg, tidseining = "month")
}
\arguments{
\item{d_ventreg}{Dataramme med oppstartsskjema.}

\item{tidseining}{Tekstring med tidseining for SPC-diagrammet som vert laga.
Sjå argumentet \code{unit} i \code{\link[lubridate:round_date]{lubridate::floor_date()}} for detaljar om
moglege verdiar.
Standardverdi er \code{"month"}.}
}
\value{
P-diagram med andel pasientar som har fått målt blodgass per \code{tidseining}.
}
\description{
Funksjonen tek inn ei dataramme med oppstartsskjema,
og lagar eit P-diagram med andel pasientar som har fått målt
blodgass per \code{tidseining}.
}
\examples{
\dontrun{
d_ventreg = hent_skjema("ventreg") |>
  dplyr::filter(lubridate::year(start_date) >= 2019)

lag_spcfigur_ki_blodgass(d_ventreg)

lag_spcfigur_ki_blodgass(d_ventreg, tidseining = "year")
}
}
