% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_spcfigur_ki_blodgass_forste_aar.R
\name{lag_spcfigur_ki_blodgass_forste_aar}
\alias{lag_spcfigur_ki_blodgass_forste_aar}
\title{Lag SPC-figur for blodgassindikator}
\usage{
lag_spcfigur_ki_blodgass_forste_aar(d_ki_superbreitt, tidseining = "month")
}
\arguments{
\item{d_ki_superbreitt}{Dataramme på superbredt format som inneholder oppstartsskjema (ventreg)
med variabelnavn r_{variabel}, og oppfølingsskjema.}

\item{tidseining}{Tekstring med tidseining for SPC-diagrammet som vert laga.
Sjå argumentet \code{unit} i \code{\link[lubridate:round_date]{lubridate::floor_date()}} for detaljar om
moglege verdiar.
Standardverdi er \code{"month"}.}
}
\value{
P-diagram med andel pasientar som har fått målt blodgass per \code{tidseining}.
}
\description{
Funksjonen tek inn ei dataramme med oppstartsskjema,
og lagar eit P-diagram med andel pasientar som har fått målt
blodgass per \code{tidseining}.
}
\examples{
\dontrun{
d_superbreitt = superbreitt_format(
  d_full_patientlist = hent_skjema("patient"),
  d_full_mce = hent_skjema("mce"),
  d_full_ventreg = hent_skjema("ventreg"),
  d_full_ventfol = hent_skjema("ventfol"),
  d_full_conclude = hent_skjema("conclude")
) |>
  dplyr::filter(
    lubridate::year(start_date) >= format(Sys.Date() - years(6), "\%Y")
  )

lag_spcfigur_ki_blodgass_forste_aar(d_superbreitt)

lag_spcfigur_ki_blodgass_forste_aar(d_superbreitt, tidseining = "year")
}
}
