% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmv-package.R
\docType{package}
\name{ltmv-package}
\alias{ltmv}
\alias{ltmv-package}
\title{ltmv: LTMV at Rapporteket}
\description{
Mainly R functions providing the shiny app at Rapporteket but may also include tools for general data analysis and presentation of LTMV registry data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rapporteket/ltmv}
  \item Report bugs at \url{https://github.com/rapporteket/ltmv/issues}
}

}
\author{
\strong{Maintainer}: Per Erik Haugedal \email{per.erik.haugedal@helse-bergen.no}

Authors:
\itemize{
  \item Per Erik Haugedal \email{per.erik.haugedal@helse-bergen.no}
  \item Endre Kvåle Evjen \email{endre.kvale.evjen@helse-bergen.no}
  \item Sergio Carracedo Huroz \email{sergio.carracedo.huroz@helse-bergen.no}
  \item Ørjan Løseth Skaar \email{orjan.loseth.skaar@helse-bergen.no}
}

}
\keyword{internal}
